First crack at porting to vmware.
I had to set the TARGET_OPTIMIZATION options in openwrt.config to
"-mtune=generic", and turn off CONFIG_MATOM and turn on CONFIG_MPENTIUM4
in trunk/target/linux/x86/config-3.12, for now.

We need to figure out how to do this "right" so as to get the right
optimization flags for each platform, _and_ (in the case of config-3.12)
keep all variants in sync except for arch-specific differences.

Review: @sberndt