Fixed the platform-specific files override issue.
Stop relying on base-files for our overrides. Instead, keep the files
in <subtarget>/files, and have rules that compose a top-level "files"
directory under trunk when building any target.

We also package that top-level files directory into the ImageBuilders, and
include those into the generated images.  In addition, however, we can pass
in an additional FILES= that will be overlaid on top of those files.