Fixes for NexCom boot and installation.
Need to bind ata-ahci as a static module, or else the kernel won't
find its own root disk when booting from the SATA disk.

Because of this, the USB boot changes so that it's always /dev/sdb2,
even when booting from USB (as long as there's a SATA disk in the box).
And because of _that_, we have to fix the TARGET_ROOTFS_PARTNAME for
NexCom/Portwell to always be /dev/sdb2.

(Unfortunately this will break the case where the box has no SATA disk
at all. In that case, we'll have to hand-patch the generated image to
use /dev/sda2 as the root partition. I tried using a UUID-based root
disk selector, but it didn't work in the NexCom.)