Added perf and valgrind to the default image, and fixed kmod-kvm for 64-bit.
Fixed the perf Makefile to support 3.12 (it was overriding CFLAGS, which
would break with the 3.12 perf source layout). Also fixed the Makefile
for "virtual" modules to also build kmod-kvm on TARGET_x64.

Finally, added both perf and valgrind to the base image as default installs.
This necessitated raising the partition size to 250MB from 200MB.