re #3412: Fix the broken flag numbering of BR_ISOLATE_MODE
The openwrt folks were moving the BR_ISOLATE_MODE flag along to higher
and higher bits as Linux kept adding more flags to the bridge code.
But they would have to remember to do this for each Linux version. We
forgot to do this when moving to 3.12, so it now collided with another
flag (BR_LEARNING), which was being set to 1 by default.

Fixed to move this openwrt-specific bit to the one of the highest bits,
so that it's unlikely to ever collide with an official Linux bit.