Re #3376 Make python use clock_nanosleep() for time.sleep()
Python uses select([]..., timeout) to implement a sleep (in floatsleep()),
but this is sensitive to the kernel clock jumping back (e.g. by ntpd). If
the clock jumps back after the sleep has started, it will sleep until the
original start time plus the timeout (which could be several hours).

Instead, I have added support for the Linux clock_nanosleep() API in librt.
With CLOCK_MONOTONIC, it is immune to such abrupt jumps.

Review: @sberndt @dnarasimhan