re #3657 @dnarasimhan @shankar@velocloud.net @sberndt
Disabling Learn2All functionality in Marvell.

Marvell driver by default sets the Learn2All bit. This causes the packets triggering new MAC leanings to be forwarded on the CPU port. When this feature is disabled, the flow looked good and we do not see the periodic (once in ~43 seconds per MAC address) intra-vlan packets on the CPU port. Also, the Learn2All feature is only needed for multi-chip switch implementation.

Assembla bug history has more details on this issue.