re #3780 @dnarasimhan @shankar@velocloud.net @sberndt
Enabling NDIS module.

More details:
1. Pantech modem has two modes of operation. In one mode, it shows 4 interface.
In another mode it shows 8 interfaces.
2. Whenever it shows up 8 interface mode (has cdc_ether interfaces), the device
gets initialized correctly. When it hits 4 interface mode, we run into trouble.
3. The reset hack that we tried brought the modem to 8 interface mode in most of
the situations. But occasionally the modem does go to 4 interface mode.

So, the problem here is, we do NOT have a reliable mechanism (currently, mode-switch does not have support to handle this modem) to bring
it to 8 interface mode consistently.

From analyzing the 4 interface mode, we observed that it shows up an NDIS
interfaces. So, we tried enabling NDIS module in openwrt and its initialization
happened correctly (USB eth interface showed up). So, irrespective of whichever
mode it shows up, we can initialize it properly.