Giant, and final(?), update of feeds.
This is getting increasingly tricky.  We have made changes to many of
the feeds packages, but any "feeds -a update" will just blow away our
local changes and refresh from the source, which is very bad.

Updated feeds/packages to 40995, and feeds/luci to 10274. feeds/packages
must NOT be pushed past 40995, because after that, they started moving
many of the key packages to github, which we can't handle.