Move Openwrt onto the "Barrier Breaker" branch (14.04), update to RC2 (42021)
From now on, we want to keep this tree as stable as possible, so we move the
subversion feed onto the Barrier Breaker (Openwrt 14.04) branch.  The current
subversion checkin version that we have sync'ed to is 42021.

We may make one more pass to pick up everything when Barrier Breaker is finally
released. Or not.