Allow generation of USB-installer and "combined" images to be turned off
Added checks on optional environment variables to disable the generation of
USB-installer or combined images.  Of course, the caller is responsible for
not passing invalid flags (e.g. turning off the combined image if also
generating a virtual disk image target, which uses that combined image).

This was the least awkward way of doing this. The right way would have been
to be able to pass in overrides to the .config file, but openwrt's Makefile
structure makes that nearly impossible.