Re #5231 Patch to dnsmasq to allow multiple configurations for DHCP/DNS
We need to be able to have different DNS configurations for secure and
guest networks.  This is a patch that came from an openwrt enhancement
request (https://dev.openwrt.org/ticket/8862) that allows this.

For now, I've also checked in the dnsmasq.init script into the velocloud
source tree (in both 1.1 and master), but that can be removed in a bit after
everyone has pulled in a new master toolchain.