added watchdog support for atom C2000; - back-ported itco watchdog from 3.18 kernel; - patched lpc_ich driver;
Unfortunatedly, lpc_ich interferes with gpio-pcu (our gpio driver).
For now, I added the watchdog mfd wrapper into out gpio-pcu driver,
because we need the BIOS fixups for Rangeley GPIO that are VC specific.
I left lpc-ich in the kernel as the VM versions may need it.
The gpio-pcu finds the PCI device first, before lpc-ich, so it wins.