Atheros wi-fi: Enable software override of firmware-limited regdomains
Some of our cards (especially the upcoming Sparklan cards, and our old
Ath9k cards from Compex for the Nexcoms) have a hard-wired regdomain in
the firmware of "0x6a" ("WORLD_1"). This forces the driver to only support
channels that are present both in WORLD_1 and whatever country we set (i.e.
an intersection).

This change sets a build flag that tells the driver to just ignore the
regdomain in the firmware, and use only the software-set regdomain.  It
also enables (for future use) DFS support, for whenever we choose to
support it (DFS-encumbered channels; needs special certification).