The great kernel-3.14 upgrade.
Updated the x64 configuration to use kernel 3.14.  The patches are all
from the openwrt svn trunk (which supports 3.14 and 3.18), and our
patches have been fixed up (thanks, sberndt).

There will be no support for this in the x86 (32-bit) platform; I will
be completely disabling that in a separate checkin of the Makefile.