Support for installing to Edge 300
Mainly, creating a first-time /etc/config/network and /etc/config/wireless
dynamically, depending on which platform we are running on.  There is a
random default /etc/config/network that gets shipped as part of the
base-files package, that we cannot nuke (yet), so the uci-defaults script
unconditionally stomps on it on first boot.

For the "real" edge500 builds (from the velocloud tree), this is unimportant.
Even though each new software update will stomp /etc/config/network back
to the default, it will again be overwritten by regenconfig.py from
vcearlyinit.

NOTE the "combined" image still refers to the edge500 device names, so it
will require hand tweaks to the grub config to boot on the edge300. But the
"installer" image will boot automatically since it uses PARTUUIDs.