Fix for hardware-dependent /etc/config/network, etc.
The uci-defaults script ran in the "wrong" place during the boot sequence
after software upgrades, and would wipe out the automatically-generated
network configuration for the actual velocloud image.

The actual problem is in how we structured vcearlyinit there, but for now,
just make the uci-defaults script a little more cautious when stomping on
the config files - look for a specific 'dummy' comment, and only stomp on
the config file if that string is present.