kmod-i2c-i801 must be installed on edge5x0, but not on edge500
When building the raw openwrt install/live images, we need to ensure that
the i2c-i801 module is installed on the edge5x0 (where it is needed), but
that it must NOT be installed on an edge500 (where it interferes with other
drivers).

The way we do this is to have the top-level Makefile rip out the lines
for kmod-i2c-i801 from openwrt.config, and then let the DEFAULT_PACKAGES
lines in the individual targets set the appropriate value for it.  For
now, only one platform (edge5x0) marks it as needed.

review @jordanr