Add real traceroute to openwrt build, fix package builds for edge500/5x0
Added a "real" traceroute (from sourceforge) to the openwrt build, and
switched the build to use that instead of the busybox traceroute.  Also
enabled traceroute6, tracepath and tracepath6 from iputils.

Fixed the configs for building kmod-i2c-ltc4266 and dmi-tool based on
the platform default packages; remove these from openwrt.config in the
top-level Makefile for each platform, and let the platform defaults be
applied.