Initial support for SuperMicro-based edge1000
vc/board will now have either "edge1000ifm" or "edge1000sm", depending
on the DMI board name.  Further, we look for a couple of possible models
of edge1000sm in edgetype: with an sfp (default for now), and with an
additional 4-port gbe with or without PoE in the future.

We use the same technique as for the edge500/300/800 to copy a proper
initial /etc/config/network into place on first boot.