Cleaned up the edge 520/540 Grub install key boot list
I've now restored the original 3 entries (Installer, Live, Fail-safe),
with a default time-out before booting the first one.

We still have internal variables called "$force520" and "$force540"
defined, so if you need to force a particular boot (for a machine whose
BIOS DMI info is incorrect), then just stop the boot, hit "e" to edit
the entry, and append "$force520" or "$force540" to the end of the line
and type "C-x" to boot that.