Ath10k: avoid 6 to 7 second wait on startup for missing calibration file
The absence of the calibration file is not an error, but the driver tries
the full firmware load sequence even if the file does not exist (causing
a 6 to 7 second delay).

Check the existence of the calibration file up front, and just skip the
firmware loading attempt if it is not present.

review @sberndt