Automatically set reboot style to "pic" for edge500 and 520/540
Adapted my original hack to Sandra's new "velocloud" base driver.
After computing the effective board type, set reboot_type to 'p'
(BOOT_CF9) for the Edge500 and the 520/540.

Note that if a reboot=x value is passed in explicitly on the kernel
command line, this is skipped (the check of "reboot_default"). This
idiom is also used in the DMI-based settings in arch/x86/kernel/reboot.c.

review @sberndt @jordanr