Bump Perl from 5.10.0 to 5.22.0
The old perl was basically broken in this build. Looks like the build
itself was a mess, and ended up linking objects with different definitions
for the same regex context struct, which caused most regex operations
to crash Perl instantly.

Fortunately, the top-of-tree Perl from openwrt 15.05+ (based on 5.22.0)
seems to be built properly, _and_ it includes a lot more useful basic
modules by default.