re#3: by default use always the qmi-proxy setup to query the modem
This allows us a clean access to the QMI device even if there's a running script
which also makes use of the QMI port. From now on, all qmicli operations out of
this script (e.g. manually calling qmicli) should be done with the explicit
'-device-open-proxy' option (or just '-p').

Note: this change requires libqmi >= 1.12.0.