re#10: actively monitor connection status and reconnect if disconnected
This change also moves from using raw qmicli Start Network commands to using the
qmi-network script, which makes it much easier to launch the connection, without
worrying about the CIDs.

During the hotplug event script we'll only generate the qmi-network profile with
the APN that we have to use. Note that the modem is not assumed to be connected
at this stage, we'll do that in the modem_service.sh run afterwards.