re#1: if modem reports being connected when detected, just setup interface
The 'PolicyMismatch' errors seem to happen when you try to connect an already
connected modem. Following this logic, if the modem was setup to auto-connect,
it may likely be already connected when we're launching our logic. Try to play
well with this, and if we detect the modem as being connected, just configure
the network interface.