re#1: if too many connection errors detected, trigger a soft or hard reset
If we get 20 consecutive checks being disconnected, we request an explicit RF
hardware reset, by putting the modem in low-power mode (which will also perform
the IMSI-detach operation) and then back in full power mode again.

If we get 100 consecutive checks being disconnected, we request a full modem
reboot.

This should help for the case where the modem gets stuck. E.g., in this case in
which it was "connected" but "not registered", the modem recovers the connection
with a soft reset:

    # tail -f /tmp/USB/USB1_ipmodem_info_collector.log
    WARNING:root:[dev=USB1]: not connected in network
    WARNING:root:[dev=USB1]: not registered in network: not-registered
    WARNING:root:[dev=USB1]: not connected in network
    WARNING:root:[dev=USB1]: not registered in network: not-registered
    WARNING:root:[dev=USB1]: not connected in network
    WARNING:root:[dev=USB1]: too many connection errors (20): requesting soft reset
    WARNING:root:[dev=USB1]: not registered in network: not-registered-searching
    WARNING:root:[dev=USB1]: not connected in network
    WARNING:root:[dev=USB1]: not connected in network
    WARNING:root:[dev=USB1]: reconnected