re#21: avoid SIGKILL-ing QMI managed modems
When the modem_service gets stopped for a given QMI modem, we shouldn't get a
SIGKILL, as that will not let us exit cleanly.

Before this patch we were sending SIGKILL to every type different than "serial",
this is to both "qmihybrid" and "ip modems". Now, we only send it for
"qmihybrid", as "ip modems" already catches signals and runs the implemented
teardown() method, if any.