re#18: IPModems,qmi: if IP address isn't set in the WWAN for a long time, hard reset
Once connected via QMI, we'll recover the IP address that we expect to see in
the WWAN interface using --wds-get-current-settings.

Then, in the monitoring loop, we'll check whether that same specific address is
set in the interface already or not. Note that udhcpc is the one responsible for
setting up the WWAN interface triggered by our ifup, as long as the WWAN was
configured in /etc/config/network.

If the IP address is not found for some time (up to 300s by default), we'll just
force a hard reset.

This situation of lacking the IP address has been noticed in a AC340U which got
stuck in a state where it would reply nicely to QMI but never to DHCP in the
WWAN. Also, this same situation is what would happen when the WWAN interface is
wrongly set to 'raw-ip' link layer protocol.

The logging in the forced hard reset case will look like:

    WARNING:root:[dev=USB4]: IP check (100.67.172.191) failed
    DEBUG:root:[dev=USB4]: Waiting for 3 secs to collect next info
    WARNING:root:[dev=USB4]: IP check (100.67.172.191) failed
    DEBUG:root:[dev=USB4]: Waiting for 3 secs to collect next info
    WARNING:root:[dev=USB4]: IP check (100.67.172.191) failed
    WARNING:root:[dev=USB4]: too many IP check errors (100): requesting hard reset
    DEBUG:root:[dev=USB4]: Waiting for 3 secs to collect next info
    WARNING:root:USB4: Received Signal 15
    DEBUG:root:[dev=USB4]: teardown
    DEBUG:root:[dev=USB4]: releasing DMS QMI client 2
    DEBUG:root:[dev=USB4]: releasing NAS QMI client 2
    DEBUG:root:[dev=USB4]: releasing WDS QMI client 1
    DEBUG:root:[dev=USB4]: setting wwan0 device down...
    WARNING:root:Exiting Program due to interrupt

The logging in a successful case will look like:

    DEBUG:root:[dev=USB3]: IP check (10.61.146.255) succeeded
    DEBUG:root:[dev=USB3]: Waiting for 3 secs to collect next info
    DEBUG:root:[dev=USB3]: IP check (10.61.146.255) succeeded
    DEBUG:root:[dev=USB3]: Waiting for 3 secs to collect next info
    DEBUG:root:[dev=USB3]: IP check (10.61.146.255) succeeded
    DEBUG:root:[dev=USB3]: Waiting for 3 secs to collect next info
    ...