re#16: hotplug: qmihybrid modems may be started from net or tty scripts
For QMI "hybrid" devices we get a network interface plus a TTY. These devices
are reported independently, and it depends on the actual modem which device
gets notified first or last.

The setup should allow launching the qmihybrid support only wehn both the net
and TTY devices have been detected.