USB key on edge 520/540 is now at /dev/sda (not sdb).
The "combined" image produced by the build has a hard-coded root path
in it, that we had previous set to /dev/sdb2 (because that's how the
original CoreBoot rom had enumerated the USB devices in the system,
including the root disk which it would always put at /dev/sda).

Now, the boot rom orders the devices in such a way that USB keys always
come first, then the built-in flash root drive, and then any mSATA
devices.  So the USB key's hard-coded device address is /dev/sda by
default.