Removed modem_functions.sh from modems package.
This was a poorly named file - it actually provided rather specific
VeloCloud functionality (to assign unique and predictable route metrics
for each interface).

In the files that used this function, I left behind a simple expression
that calculated basically the same value for USB* interfaces, so that
this module can be self-contained in a raw openwrt image.

Notify @gopakumar