Generate OVA packages for vc-vmdk images
A very fiddly process, which is explained in README-ovf.txt in
the target/linux/x64/image directory.  I'm sure this won't be the
final version, and changes will be needed for the OVF after feedback.

Include a binary package downloaded from VMware for the virtual disk
manager (vmware-vdiskmanager).

notify @kangwarn @sberndt