re#40: setup new 'modem_apn' script to manage user-provided APNs
The script will now store the modem 'profile' in a location which is persistent
across reboots, not under /tmp.

Also, the profile is only automatically created if there isn't one already set.
This makes it possible to keep the old automatic logic around, and still allow
user-provided updates later on.

If the service is flagged as started by the hotplug scripts, then we'll restart
the service, even if the actual process isn't around any more.

Only for IPModems (i.e. WWAN based) for now.