re#19, re#20: modems: consistent start/stop/restart actions
The 'stop' operation will try to kill the process several times, including
different signal types (first SIGTERM, then SIGINT, then SIGHUP and as a last
option, SIGKILL). Once each of these signals is sent, we check whether the
process is still running, and if it isn't we remove the PID file, so that the
'status' option works correctly (re#19).

We also don't try to 'start' the modem script if the modem is unavailable or
is of an unknown type.

Last but not least, some of the operations that are done unconditionally, like
some route removals in the 'stop' actions are updated to ignore errors reported
in stdout (re#20).

See the following example sequence:

    root@vc:~# /etc/modems/modem_service.sh USB4 start
    USB4: starting...
    USB4: starting WWAN modem support
    USB4: started modem qmiwwan script sucessfully

    root@vc:~# /etc/modems/modem_service.sh USB4 status
    USB4: modem script running

    root@vc:~# /etc/modems/modem_service.sh USB4 stop
    USB4: stopping...
    USB4: stopping process with pid 22280 (SIGTERM)
    USB4: process with pid 22280 now stopped

    root@vc:~# /etc/modems/modem_service.sh USB4 status
    USB4: modem script not running