Added mce-inject and kmod-mce-inject, updated mcelog to latest
Updated mcelog to the latest GIT version, and added a --no-dmi option
(because our edge500's and edge5x0's do not populate a DIMM section, or
fill it with an unexpected (to mcelog) format for the Bank information).

Added mce-inject, and kmod-mce-inject.  The latter is NOT auto-loaded; it
must be explicitly loaded by insmod.  This is for testing MCE handling
and logging. Data file format examples can be seen at the mce-test
repository at http://git.kernel.org/cgit/utils/cpu/mce/mce-test.git .

notify @jordanr @sberndt