Small improvements to VeloCloud Edge ovf.
Attach the virtual disk to SCSI (paravirtual) bus rather than IDE bus.
Delete one of the IDE buses (we don't need two). Added a <ProductSection>.
Fixed OVF generation script (substitution of variables in ovf template).

This was in response to a customer report that they couldn't change the
size of a virtual disk (the spinner was grayed out in the UI).

notify @varun856283