Patched .svn files to point to svn.mein.io
svn.openwrt.org is finally dead. They are forcing everyone to go to the
git repository (git.openwrt.org) instead, but that will require a totally
different type of organization (and probably a new repository).

In the meantime, one of the openwrt owners (John Crispin) has set up a
temporary svn server at svn.mein.io, so that we can migrate gracefully.

The first step is pointing the .svn files at the temporary repository,
so that we see what changes we have made.  I will also move us as close
as possible to OpenWrt 15.05.1 (Chaos Calmer) for most of the branches.

Work on a new toolchain based on the git repository will start soon.