re #11926 workaround for dnsmasq infinite loop on large DHCP request packet
recvmsg() with MSG_PEEK|MSG_TRUNC seems to be broken in our Edge kernel;
if called with a buffer smaller than the packet, instead of returning
the actual size of the message (and the trunc flag), it returns -1
with errno==EFAULT (even though all the parameters are correct and in
the process space).  This causes dnsmasq to go into an infinite loop
trying to read that packet.

Work around this by sizing the initial DHCP packet receive buffer to
the maximum UDP packet size (1500), so that it never triggers the above
failure condition.

(Checked with the latest 3.14.66 kernel; it hasn't fixed this problem.)