re #12092 Add out-of-tree ixgbe driver to replace the kernel's own ixgbe
Our kernel is so old that its included ixgbe driver doesn't support any
of Intel's recent adapters like X550.

Add the Intel-distributed "portable" driver as an out-of-tree driver,
and disable the ixgbe in the kernel.

WARNING: this still has the issue that it barfs on unsupported SFP+
modules, but now it's worse because when it sees one of these, it actually
unregisters the ethX device immediately, which is really drastic.

Fixes coming for these problems soon.