re #12092 Support for the Supermicro edge800 in the startup scripts
Fixed the edgetype and uci-defaults scripts to support the new Xeon-D
SuperMicro edge800.  Also fixed the support for the "enhanced" edge 1000
(supermicro) with the QAT card and 6 GbE + 2 SFP+ ports.

This allows the edge to come up with the right network configuration
automatically for the platform. (raw openwrt image only - the production
image from the velocloud repository has a different mechanism for insuring
this).