Workaround for CONFIG_DEBUG_KERNEL being disabled even though selected
We actually enable CONFIG_DEBUG_KERNEL in the kernel config-3.14, but
openwrt has its own overriding option CONFIG_KERNEL_DEBUG_KERNEL, and
also has a bug that if all of the options that select it are turned off,
it ignores an explicit setting of this option, and worse, disables the
actual config option in the kernel.

So we select one of the selecting options for it, so that it stays
selected, and results in CONFIG_DEBUG_KERNEL being enabled properly at
the kernel level.