Preliminary checkin for updatable virtual edges
This checkin changes the type of images built for all virtual edges
(except EC2 images).

Now, instead of a classic "combined" image (2 DOS partitions - boot+root),
it directly generates a 6-partition GPT setup, just like for newer hardware
edges than the 500.  There is a grub partition, a boot partition, 3 roots,
and a velocloud partition.

This velocloud partition also allows logs and cores to be saved across
boots, just like for the hardware edge.

The Disk Image is now 8 GB (3 1-GB roots, and a 4.2 GB /velocloud partition).
The OVF for ESX deployment now has 2 CPUs and 4 GB of memory (we can change
these after deploying the OVF, before booting it up), and 5 GE ports - 2
switched and 3 routed, called GE1..GE5.  Up to 3 more can be added in the
field after deployment, if needed.

NOTE that we did not do this for the EC2 image type, since that would
require a separate development effort to chain bootloaders (from AWS's
bootloader to something we put in).

NOTE also that during the first boot, there will be a noticeable delay as
the /velocloud partition is expanded to fill the partition. To speed up the
build, we create the partition of the right size, but only a tiny filesystem
within it, and on first boot, we resize2fs the partition to the full size.