modems: integrate ModemManager support
By default the new logic will launch the ModemManager support as soon as a new
device is found, except for when the device VID/PID is added in the blacklist.
So, if the device is blacklisted in ModemManager, we'll try to take control of
it ourselves with the legacy "TTYModem"/"IPModem" implementations.

If the device is really managed by ModemManager, we'll launch the new "MMModem"
implementation, which will wait for ModemManager to expose the device in DBus,
and once that is done, it will try to get it enabled, registered and connected.

The MMModem implementation relies on the "modem_apn.sh" script to gather the
APN required in the connection, if any. For 3GPP2 devices which don't require
an APN, this can be left blank.

Once the device is in connected mode, the MMModem implementation will activate
the connection using the method requested by ModemManager for this device:
   * PPP: For modems where the data port is a TTY, pppd will be launched.
   * Static: For modems where we can gather the IP settings via control
     commands, the static IP configuration will be set for the WWAN interface.
   * Dynamic: For modems with a built-in DHCP server, DHCP will be used to
     request and setup the IP settings.