re #12568 fix for ext4 assert in VC_VMDK kernel
See the patch description for what is wrong (and the symptom). I ran
into this with my vc-vmdk kernel, which leads to a spiral of corruption.

(Anupam: this could be the root cause of your occasional inability to
reboot; when this happens, any attempt to sync or reboot just hangs.)

notify @sberndt @anupamrai