Fix the console parameters for virtual edges in generated image
Turns out that "console" in a VM (both VMware and KVM) translates to
tty1, not tty0.

Also, when specifying multiple console= options on the kernel command
line, it's the _last_ one that gets used as the standard I/O tty for
prompts from the preinit scripts.  image/Makefile hard-codes "tty0" as
the tty for "console", and always specifies it _before_ any serial
console.

The right fix here is to have an additional parameter that is the
"preferred" console, and make sure it's the last (and also make it
configurable), so that we don't have to do this hack, and also so that
we can make this work on the 800 and 1000 as well (where we want the
VGA console, for now, to be the "primary" console..)