VMware OVF: make GE1/GE2 not connect at power on, and GE3/GE4 connect
By default, when the OVF is deployed, some of the interfaces are set
to connect at power-on, while the others are not (the effect of that
is the same as if on a real edge, you disconnect the cable).

But we had wonky defaults: GE1/GE2 (the "switched" interfaces) were
connected at power-on by default (bad), and GE3/GE4 (default routed
interfaces) were _not_ connected at power-on.

Changed this around, so that a VM doesn't immediately start responding
incorrectly to things like DHCP requests, unless this is really
intended.