re #12833 ensure that all VM edge filesystems are journaled ext4
genext2fs doesn't create journaled filesystems. Which didn't matter
for gen_image_velocloud.sh, since that was just building an installer;
the installer itself executed mkfs.ext4 which does the right thing.

Here, we are actually building the full VM image, so we have to convert
all of the genext2fs-generated filesystems to "ext4" as best as we can,
which means tune2fs calls to add journaling, and a host of other ext4
features.

For the "user" partition itself ("/velocloud"), we now use a real
mkfs.ext4 call on a zeroed-out file of the right length (16MB for the
initial image), so that it creates a journal, *and* sets all the correct
options for being able to do online resize2fs (which the above tune2fs
sequence does not!).  This makes the first-time expansion of /velocloud
lightning-fast as a happy side-effect.

notify @connors this will be in the next toolchain(s).