re #3960 #12359 hard reset: boot directly from grub to root 0
Use the new functionality from sberndt to capture the hard reset button
state in grub, and use that to launch root 0 directly (i.e. launch the
"Factory Default" menu as the default).

Note: this is only part 1 of this (for new boxes).  For part 2, we want
to be able to update the grub on existing edge500 systems in the field,
and that will come later.

review @jordanr @sberndt