modemmanager: new package built from git
The new ModemManager support in OpenWRT is based on the new 'ReportKernelEvent'
method in the Manager DBus interface, which allows reporting udev-like kernel
events via DBus, in our case coming from hotplug2 scripts.

The package installs hotplug scripts for tty and net devices (cdc-wdm devices
in MBIM and QMI are reported along with the net ones). For each device reported
within hotplug, the scripts will report kernel events to ModemManager, unless
the device VID/PID is explicitly listed in a "blacklist".

The package also installs a service file so that ModemManager is started on
system boot.

NOTE: this package requires 'xsltproc' and 'intltool' installed in the HOST.