re #13666 formal Quagga and DPDK packages for the gateway
Build .deb packages for quagga and dpdk (dkms) for the gateway.
These are built from the "host" build directories, though they are
really not building host packages for the SDK - they're building .deb
packages for the gateway.

See the README.versions file in the main and the host package directories
for these products, for special instructions on how to keep the builds
in sync, and bump up version/release strings as needed for the gateway.

review @gopakumar @vipinkr @jordanr