re #14442 Attempted fix for ubusd/ubus 100%-cpu loops
Catch the situation where the unix-domain socket read fails with an
error that is not EAGAIN/EINTR, and terminate the resultant loop.
I have picked up the most recent ubus and libubox releases for this,
and applied this fix as an increment on top of the other fixes already
available in the newer versions.

Hopefully this fixes the situation, but it needs beaucoup testing.

review @connors @arun.srinivasan