re #14658 Add drivers for virtual platforms to the kdump kernel
The kdump kernel was missing drivers required for VMware, KVM and Xen.
As a result, it couldn't find the root disk in many cases (if it was
pvscsi or scsi_virtio, for example), and just hung.

Added several virtual drivers.  The resultant kdump kernel boots on both
virtual (e.g. VMware ESX) and hardware (e.g. edge500) platforms.

review @jordanr