re #14648 Add a new interrupt type to look for: RTE_ETH_EVENT_VF_MBX_SYNC
Ensures that when VF mailbox is out of sync with host PF, we do a stop/start
to resume normal operation.

On the host side, we better see the kernel message:  VF Reset msg received
from vf 0 on a PF flap.

Adding back for Jordan.  His local repository was a bit out of sync.

review @jordanr @gopakumar