Re #15179 make peer address check specific to a view/vrf
Quagga has a global iflist which is referred to across
all views in BGPd. That was causing issues with a peer-address
sanity check, as it would match with IP subnet in other view.

Fix is to also match bgp instance name with the instance in
the interface before going ahead with the check.