Dolphin network bring-up: igb-eeprom and IGB driver changes
Added a new VC_ID for Dolphin in the IGB eeprom, along with the right
MDIO parameters for it, and a small range of MAC addresses for the
dolphin proto (-B edge510 -u 510-prod:<serialno>).

In the IGB driver (igb_vc_probe), look for the new VC_ID, and bypass
probing for the switches, etc. and just do an ordinary probe for
supported phys.

review @sberndt