Enable debug build of edge quagga
The edge bgpd and ospfd will be built with symdebug, just like edged,
so that we can make sense of their cores.

The gateway bgpd is still TBD - we will have to see if it's OK to enable
-g -O0 for it and still get reasonable performance, or whether we'll have
to experiment with -O3 -g or something (less useful cores and traces).

review @vipinkr @gopakumar