toolchain: fix default dhcp.script included in netifd
Fix the network route metric setting in set_classless_routes(), by
passing the metric as first argument instead of as last argument (as we
don't know exactly the number of arguments being passed in the
'staticroutes' and 'msstaticroutes' variables).

Note that this script is only used when the toolchain is built alone; it
will be overridden in the velocloud tree, which has its own way of
computing the metric of each network interface.