modems,tty bug 86: avoid processing multiple events per TTY
For each TTY we get two hotplug events, one for the 'usb-serial'
subsystem and one for the 'tty' subsystem:

    HOTPLUG:Wed Feb  8 11:48:11 UTC 2017:MODEM1: usb: ttyUSB0
    DEVNAME=ttyUSB0 USER=root ACTION=remove SHLVL=2 HOME=/ SEQNUM=1984 HOTPLUG_TYPE=tty MAJOR=188 DEVPATH=/devices/pci0000:00/0000:00:16.0/usb1/1-1/1-1.4/1-1.4:1.0/ttyUSB0/tty/ttyUSB0 DEVICENAME=ttyUSB0 LOGNAME=root TERM=linux SUBSYSTEM=tty BOOT_IMAGE=/boot/vmlinuz PATH=/bin:/sbin:/usr/bin:/usr/sbin MINOR=0 panic_timeout=5 PWD=/ PYTHONPATH=/opt/vc/lib/python

    HOTPLUG:Wed Feb  8 11:48:13 UTC 2017:MODEM1: usb: ttyUSB0
    USER=root ACTION=remove SHLVL=2 HOME=/ SEQNUM=1985 HOTPLUG_TYPE=tty DEVPATH=/devices/pci0000:00/0000:00:16.0/usb1/1-1/1-1.4/1-1.4:1.0/ttyUSB0 DEVICENAME=ttyUSB0 LOGNAME=root TERM=linux SUBSYSTEM=usb-serial BOOT_IMAGE=/boot/vmlinuz PATH=/bin:/sbin:/usr/bin:/usr/sbin panic_timeout=5 PWD=/ PYTHONPATH=/opt/vc/lib/python

Until now we were ignoring the 'tty' subsystem event because the logic
loading the 'vid:pid' string would only work with the DEVPATH of the
device in the 'usb-serial' subsystem. For the 'tty' subsystem event we
would end up logging this misleading message before ignoring it:

    Wed Feb  8 10:39:33 2017 user.notice usbmodems: MODEM1: usb: ttyUSB2: Ignoring tty device, no product: DEVICENAME:ttyUSB2

Instead of doing this, we now provide a common method to load vid:pid
based on traversing the DEVPATH until the physical USB device DEVPATH is
found (the one with idVendor and idProduct files). This method works
with both DEVPATHs from both events.

And then, we filter out events in the TTY hotplug scripts which don't
have a DEVNAME parameter, which is the same logic applied in the
ModemManager hotplug scripts.