modems,qmi bug 89: try QMI LLP change in the plugin, not in hotplug
The logic that does the LLP change for QMI devices sets up a loop with
several retries, so that the Netgear AC341/AC340 are correctly switched
from raw-ip to 802.3. But doing that in the hotplug script ends up
blocking all other hotplug scripts for a long time.

Try to avoid that, by launching the LLP change in the QMI plugin
implementation, which runs as a service spawned from the hotplug script,
and therefore doesn't block other hotplug scripts.