More Dolphin tweaks for reboot/reset support
Fix the velocloud-vc driver to also enable reboot=p automatically for
Dolphin.  Requires PIC v5 or later.

Also invoke a platform-specific script in vcearlyinit that performs a
number of initialization steps on Dolphin:
* Disable the PIC watchdog timer for now (until supported in procd)
* print informative messages about the PIC version and restart reasons

There is also a commented-out statement to set the "cold restart" type
in PIC to power off and on, but we won't enable it for now, until we
shake out as many initialization problems as we can early on.  If
things still don't stabilize by production time, we may enable this.

review @sberndt @jordanr