re #16760 add kernel param and sysctl to reboot box on RCU stall
virtio in vc-kvm-guest running in AT&T's uCPE sometimes stalls in
rcu/tree.c when killing edged and unloading the DPDK drivers.

If this happens, we can now force a reboot (really, a panic via
BUG_ON, which shows up as a kernel panic like this:

[  311.092832] kernel BUG at kernel/rcu/tree.c:984!
[  311.097988] invalid opcode: 0000 [#1] SMP

The key here is that if it BUGs in rcu/tree.c, then it's because of
an RCU stall. We will only set this sysctl in the AT&T environment.

review @gopakumar @sberndt