modems, bug 97: on startup make sure the network is reloaded
When a power outage occurs the Edge will go offline with the
interfaces configured in /etc/config/network.

When the Edge comes back online, the modem scripts will clear the USBX
configurations from /etc/config/network, but they were not doing a
network reload, so if netifd had already parsed those configurations,
the removal wasn't being taken into account.

When the modem is processed and the hotplug scripts re-create the
/etc/config/network entry and do a reload, netifd did NOT see any
change in the USBX settings due to the missing previous reload, and
therefore didn't run the DHCP client on the network interface.

This fix provides the missing network reload when we do the
/etc/config/network file cleanup on startup.