Re #17757 BGP routes from PE is not installed on the OPG
This issue is specifically about having the local-IP in customer A
matching the neighbor IP in customer B, which causes the martian
next-hop sanity to fail.

And thats because local-IPs are stored in a global hash table.

Fix is to make this address hash per BGP view/vrf.

Review @arun.srinivasan @rnavanee @dataplane